requirements.txt
adbutils==1.2.9
av==10.0.0
customtkinter==5.1.3
numpy==1.24.3
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
requests==2.30.0
tesserocr==2.6.0
tk==0.1.0
